done strash

+++ fraig/src/cir/cirFraig.cpp
@@ -41,8 +41,9 @@ CirMgr::strash()
       CirGate* g = _dfsList[i];
       if(g->getType() != AIG_GATE) continue;
       if(g->getfinSize() != 2) return; //error
-      HashKey key(g->getInputWithInv(0), (size_t)g->getInputWithInv(1));
-      
+
+      HashKey key(g->getInputWithInv(0), g->getInputWithInv(1));
+      // cerr << g->getInputWithInv(0) << "," << g->getInputWithInv(1) << endl;
       if (map.query(key, g)) { // collision happens!
       	CirGate *tmp = _dfsList[i]; // g is old value, and tmp is new value now
       	// replace tmp with g
